Analyze structure of a function body

Dissects a given function body (technically, any call) to determine if forbidden functions with, return or FU_... occur anywhere, and to determine how the state y and parameter p are indexed. Functions FU_... are used internally to represent the state variables and therefore may not be used for other purposes. It does these checks by going recursively through the call tree. This function assumes that y and p are the formal arguments representing the state and parameter vectors.

analyzeBody(
  callExpr,
  out = c(y.name = FALSE, y.nr = FALSE, y.comp = FALSE, y.vec = FALSE, y.brc = FALSE,
    p.name = FALSE, p.nr = FALSE, p.comp = FALSE, p.vec = FALSE, p.brc = FALSE,
    theta.name = FALSE, theta.nr = FALSE, theta.comp = FALSE, theta.vec = FALSE,
    theta.brc = FALSE, eta.name = FALSE, eta.nr = FALSE, eta.comp = FALSE, eta.vec =
    FALSE, eta.brc = FALSE, eps.name = FALSE, eps.nr = FALSE, eps.comp = FALSE, eps.vec =
    FALSE, eps.brc = FALSE, with = FALSE, return = FALSE, FU = FALSE)
)

Arguments

callExpr

A call, for example a function body.

out

Boolean vector storing the output indicators. Initialized to FALSE and this should not be changed.

Value

Boolean vector of the same format as out, where TRUE means the corresponding element is present in callExpr, and FALSE means it is not. The elements are y.name, y.nr and y.comp indicating whether the vector y

is indexed by names (strings), numbers or by more complicated expressions (e.g. arithmetic expressions or functions). The element y.vec indicates whether y is referred to as a vector. The element y.brc indicates whether elements of y are referred to with a single bracket ('[]') or '$'. More than one of these may be true, if multiple reference methods are used in callExpr. The elements p.name, p.nr, p.comp and p.vec do the same for p, and likewise for theta, eta and eps. Elements with, return and FU indicate the presence of the associated functions. If callExpr contains illegal components (anything other than a call, pairlist, atomic or name), then execution is stopped.

Author

Martijn van Noort