Calculate output in case of symbolic derivatives

Calculates the variations in case of symbolic derivatives, for a given output and parameter vector, both in variational form, that is, including derivatives. For the output vector these are derivatives with respect to the derived parameters (p). For the parameter vector these are derivatives with respect to the base parameters (theta, eta, eps), denoted here as v. This function combines these to obtain derivatives of the output with respect to the base parameters. This is an applicaton of the chain rule for derivatives: dy_i/dv = sum_j(dy_i/dp_j dp_j/dv) and d^2y_i/(dv dw) = sum_j,k(d^2y_i/(dp_j dp_k) dp_j/dv dp_k/dv) + sum_j(dy_i/dp_j d^2p_j/(dv dw)).

calcVarSymb(doutput, dpout, vartheta, vareta, vareps, secOrd)

Arguments

doutput

data frame containing values of derivatives of order 0, 1 and optionally 2 of the output vector, each row in the format of the output of the function doutput, see createDerivsSymb, and each row corresponding to a time point. The column names should reflect the component and derivative, as in the output of the function doutput.

dpout

named numeric list containing values of derivatives of order 0, 1 and optionally 2 of the parameter vector, in the format of the output of the function dp, see createDerivsSymb.

vartheta

Vector of names of population parameters for which variations are to be calculated. Should be a subset of names(theta), or NULL for none.

vareta

Vector of names of individual parameters for which variations are to be calculated. Should be a subset of names(eta), or NULL for none.

vareps

Vector of names of residual parameters for which variations are to be calculated. Should be a subset of names(eps), or NULL for none.

secOrd

TRUE if second order derivatives have to be computed, FALSE if not.

Value

A vector with derivatives of order 0, 1, 2 of the state variables with respect to base parameters (theta, eta and eps). The elements are named <nam>.<i>, where i is the sequence number of the component of y, and nam is of the form 'y', 'dy_d<v1>' or 'd2y_d<v1>_<v2>'.