Check formatting of model components for symbolic manipulation

Checks whether model components (model, p, init, output) are formatted correctly for symbolic manipulation. The following checks are performed:

Each component has the correct formals.

The components do not contain "with" or "return" statements, or functions starting with "FU_". The latter are used to represent the state variables in internal calculations and therefore may not be used for other purposes.

The components do not refer to y, p, theta, eta or eps as vector or indexed via ".".

The components refer to elements of y, p, theta, eta or eps in the same way, i.e., either by name or number.

The vectors theta, eta and eps are consistent with this (i.e. whether they name their components or not).

The outputs of the functions init and p are consistent with this (i.e. whether they name their outputs or not).

checkModelDef(model, p, init, output, theta, eta, eps, showWarn)

Arguments

model

Function(t, y, p) of time, state variables and parameters, specifying the differential equations. This function should return the numeric vector dy/dt.

p

Function(theta, eta) of population and individual parameters specifying the model parameters. This function should return a named numeric vector.

init

Function(p) of parameters specifying the initial state (at time 0) of the model.

output

Function(y, p, eps) of state variables, parameters and residual errors specifying the model outputs. This function should return a numeric vector.

theta

Population parameter values, as named numeric vector. May be NULL if none are needed.

eta

Individual parameter values, as named numeric vector. May be NULL if none are needed.

eps

Residual parameter values, as named numeric vector. May be NULL if none are needed.

showWarn

TRUE if warnings about missing random parameters are to be shown, FALSE if not. The latter is of use in case only the structural model is of interest.

Value

TRUE if the model components pass all tests, FALSE if not. In the latter case, error messages will be displayed. All encountered errors are shown, but if the formals are incorrect, the other checks are not performed. In the first case, warning messages may be displayed.

Note

theta, eta and eps are only used to generate some output values of the functions p and init, so that their naming can be checked, and to check whether they have names or not. Their specific values usually do not matter (but see next note).

These checks are meant to catch the most common mistakes but are not completely failsafe. For example, to check consistency in naming they determine if p and init provide named or unnamed return values for the given theta and eta. This may fail to detect errors in case the naming depends on the parameter values. Another example is that the validity or consistency of names or indices is not checked (e.g., one model component may use different names than another, or indices can be negative).

Author

Martijn van Noort