Check variance-covariance matrices and return them in a standard format

Performs several checks on the variance-covariance matrix inputs of the methods that calculate the FIM, and return them in a standard format. Can also be used to standardize the input to normalizeFim. For the inputs omega and sigma it is checked that they are matrices that are square, symmetric, positive definite and with matching row and column names. Those names are set equal if necessary. For the inputs varomega and varsigma it is checked that they are symmetric matrices, or single values from which symmetric matrices are constructed. Their row and column names are set to match those of omega and sigma.

checkSetCovMat(omega, sigma, varomega, varsigma, caller)

Arguments

omega

The variance-covariance matrix of the individual effects, provided as a numeric matrix.

sigma

The variance-covariance matrix of the residual effects, provided as a numeric matrix.

varomega

Indicators which elements of omega should be considered as variable, provided as a boolean matrix or as a single boolean or NULL. In case of a single boolean, it is assumed that this single value applies to every matrix element (also the off-diagonal ones). The value NULL means all nonzero elements of omega are TRUE, and the others are FALSE.

varsigma

Indicators which elements of sigma should be considered as variable, provided as a boolean matrix or as a single boolean or NULL. In case of a single boolean, it is assumed that this single value applies to every matrix element (also the off-diagonal ones). The value NULL means all nonzero elements of sigma are TRUE, and the others are FALSE.

caller

String containing the signature of the calling function, for use in error messages. A typical value would be 'FisherInfo::calcFimFromMatrix'.

Value

NULL if there were errors. In this case an error message will be printed. Otherwise a list of four elements. The first two are the variance-covariance matrices, i.e. omega and sigma. The second pair are the indicator matrices varomega and varsigma. All four are adapted from input as described above.

Details

Superfluous rows and columns are removed from these matrices. That is, any row (and corresponding column) with only zero variances and covariances in omega and only FALSE in varomega will be removed from both matrices, as it would not affect the outcome. Likewise for sigma and varsigma. The resulting matrices omega and sigma should not both have dimension zero.

Author

Martijn van Noort