Check that the matrix df contains all required derivatives and return their column names

The FIM calculation requires all first derivatives "dy_d<p>" and all second derivatives "d2y_dp_q" or "d2y_dq_p", where p can be any element of vartheta or any rowname of omega or sigma, and q can be any rowname of omega or sigma.

checkSetDerivCols(df, omega, sigma, vartheta, caller)

Arguments

df

Data frame of first and second order variations as generated by calcVariations or calcVariationsFim, that is, with columns 't', 'i', 'y', 'dy_d<v1>' and 'd2y_d<v1>_<v2>', where variables v1 and v2 are replaced by names.

omega

The variance-covariance matrix of the individual effects, provided as a numeric matrix.

sigma

The variance-covariance matrix of the residual effects, provided as a numeric matrix.

vartheta

Vector of names of population parameters for which derivatives are to be calculated. Should be a subset of the names appearing in the variational matrix, or NULL for none.

caller

String containing the signature of the calling function, for use in error messages. A typical value would be 'FisherInfo::calcFimFromMatrix'.

Value

NULL if there were errors. In this case an error message will be printed. Otherwise a list of five elements, named "theta", "eta", "eps", "thetaEta" and "thetaEps" containing the column names in df of the derivatives of the indicated type.

Details

This function checks that the matrix df contains all these derivatives, and returns their names in df. It also checks that the derivative names are returned in canonical order, i.e., corresponding to the order in omega and sigma.

Author

Martijn van Noort