Create grep strings for derivatives.

Creates grep strings for selecting the given derivatives from the output of Deriv. To this end, the codes provided in derivCodes are translated to match the names of elements of this output. The grep strings take into account that these names can be different, depending on simplifications of the output. That is, output vectors equal to 0 are sometimes reduced to single elements, and corresponding parts of their name are suppressed. Also, derivatives may be wholly or partially suppressed.

createDerivGreps(
  derivCodes,
  compNames,
  varName,
  derivSep = "\\.",
  typeSep = "_",
  outnameSep = "_"
)

Arguments

derivCodes

Vector of strings containing the derivative codes. Elements have the form 0 for the 0-th derivative, 1<derivSep><deriv> for the 1-st derivatives, and 2<derivSep><deriv><derivSep><deriv> for the second derivatives. Here <deriv> codes a derivative as <type><typeSep><variable>, where <type> is a derivative type, such as "theta" or "eta", and <variable> is a variable (from the list of variables of the given type), such as "TVCL".

compNames

Vector of strings specifying the names of the output components of the 0-th derivative, or NULL if they have no names.

varName

String specifying the variable name to be used in the output.

derivSep

String specifying the character(s) separating the various parts of an element of derivCodes.

typeSep

String specifying the character(s) separating the type and variable in elements of derivCodes.

outnameSep

String specifying the character(s) separating the various parts of the derivative specification in the names of the output.

Value

The grep strings, as a named vector of strings of the same length as derivCodes. The names of its elements are <varName><outnameSep><deriv1><outnameSep><deriv2>, where the derivatives are included as needed.

Author

Martijn van Noort