Generate derivatives of model components symbolically

Calculates first and optionally second derivatives of the model, initialization, output and parameter functions with respect to relevant parameters, in a symbolic way.

createDerivsSymb(
  model,
  p,
  init,
  output,
  theta,
  eta,
  eps,
  vartheta,
  vareta,
  vareps,
  secOrd
)

Arguments

model

Function(t, y, p) of time, state variables and parameters, specifying the differential equations. This function should return the numeric vector dy/dt.

p

Function(theta, eta) of population and individual parameters specifying the model parameters. This function should return a named numeric vector.

init

Function(p) of parameters specifying the initial state (at time 0) of the model.

output

Function(y, p, eps) of state variables, parameters and residual errors specifying the model outputs. This function should return a numeric vector.

theta

Population parameter values, as named numeric vector. May be NULL if none are needed.

eta

Individual parameter values, as named numeric vector. May be NULL if none are needed.

eps

Residual parameter values, as named numeric vector. May be NULL if none are needed.

vartheta

Vector of names of population parameters for which variations are to be calculated. Should be a subset of names(theta), or NULL for none.

vareta

Vector of names of individual parameters for which variations are to be calculated. Should be a subset of names(eta), or NULL for none.

vareps

Vector of names of residual parameters for which variations are to be calculated. Should be a subset of names(eps), or NULL for none.

secOrd

TRUE if second order derivatives have to be computed, FALSE if not.

Value

A list of four functions, containing the variational versions of model, p, init and output. These functions take same arguments as their non-variational counterparts, but with a different meaning for the argument y, namely this contains the state variables and their derivatives wrt parameters. The output of p is a named list of named numeric vectors, with one element for each derivative. The names of the elements have the format p(_<deriv>(_<deriv>)), where <deriv> codes the derivative and is optional, as indicated by the brackets. Each element has the same length as the output of the non-variational version of p, and its components have the same names. The output of init and model is a named numeric vector of values of the state variables and their derivatives. The names have the format y(_<deriv>(_<deriv>)).<state>, where <deriv> codes the derivative and is optional, and <state> codes the state. The output of output is a named numeric vector of values of the output variables and their derivatives. The names have the format y(_<deriv>(_<deriv>)).<state>, where <deriv> codes the derivative and is optional, and <state> codes the state or (if the state has no name) its index.

Author

Martijn van Noort