Calculate Fisher Information Matrix

This is the workhorse for all FIM calculation methods. The difference with calcFimFromMatrix is that it does not check the validity of the variance-covariance matrices, and assumes that they are in the appropriate format. It calculates the Fisher Information Matrix (FIM) from the given variational matrix, that is, calculates minus the expectation of the second order derivative matrix of the log-likelihood at the given times with respect to the model parameters. The times are given in the variational matrix, which contains first and second order derivatives of the output with respect to certain parameters, at certain times, for a particular individual.

doCalcFim(df, omega, sigma, vartheta, varomega, varsigma, caller)

Arguments

df

Data frame of first and second order variations as generated by calcVariations or calcVariationsFim, that is, with columns 't', 'i', 'y', 'dy_d<v1>' and 'd2y_d<v1>_<v2>', where variables v1 and v2 are replaced by names. This should not be normalized. If it is, it is denormalized.

omega

Symmetric, positive definite, numeric variance-covariance matrix of individual parameters. Columns and rows should be named, and those names should be the same.

sigma

Symmetric, positive definite, numeric variance-covariance matrix of residual parameters. Columns and rows should be named, and those names should be the same.

vartheta

Vector of names of population parameters for which derivatives are to be calculated. Should be a subset of the names appearing in the variational matrix, or NULL for none.

varomega

Symmetric boolean matrix of the same dimension as omega, with the same row and column names. The value TRUE means the derivative of the corresponding element should be included in the FIM, FALSE means not.

varsigma

Symmetric boolean matrix of the same dimension as sigma, with the same row and column names. The value TRUE means the derivative of the corresponding element should be included in the FIM, FALSE means not.

caller

String containing the signature of the calling function, for use in error messages. A typical value would be 'FisherInfo::calcFimFromMatrix'.

Value

The FIM, as a square matrix with element (i,j) equal to -E(d^2 L/d<vi><vj>), where L is the log likelihood and <vi> is the i-th parameter. The row and column names of the matrix are those of the variable thetas, omegas and sigmas, as specified by vartheta, varomega and varsigma. The FIM is not normalized and the 'thetaNormalized', 'omegaNormalized' and 'sigmaNormalized' attributes of this matrix are set to named vectors of FALSE, where the names are the names of the thetas, and the column names of omegas and sigmas as specified by df, omega and sigma, respectively. The attributes 'theta', 'omega' and 'sigma' store the values of all parameters as named vectors. They form a superset of the variable parameters. The 'type' attribute is set to 'FIM' The function displays an error and returns NULL if the input is incorrectly formatted or the variance matrix cannot be inverted.

Author

Martijn van Noort