Get the column parameter values of a FIM

Gets the values of the parameters corresponding to the columns of the FIM. This is a subset of getAllParsFim corresponding to the columns present in the FIM.

getParsFim(fim)

Arguments

fim

The FIM.

Value

A named list of three named numeric vectors containing the theta's, omega's and sigma's used to create the FIM, and present as columns of the FIM, in the order of appearance in the FIM. The list elements are named "theta", "omega" and "sigma", respectively. The vector elements have the names of the parameters, where off-diagonal elements of omega and sigma are named "(<par1>,<par2>)". The function displays an error and returns NULL if fim is not a valid FIM.

See also

Author

Martijn van Noort