Plot of normalized sensitivities and aliasing scores

Generates several plots for normalized sensitivities and/or aliasing scores, as computed by calcAliasingScoresFromMatrix or calcAliasingScoresFromModel.

plotAliasing(aliasing, elt = NULL, vars = NULL)

Arguments

aliasing

Named list of aliasing score output as produced by calcAliasingScoresFromMatrix or calcAliasingScoresFromModel.

elt

List with elements "S", "P", "A" and/or "T" indicating which plot(s) should be made. NULL can be used as shorthand for all.

vars

Vector of names of parameters to be included in the plots. If NULL (default) or missing, all parameters in aliasing are used.

Value

A named list of plots with the same names as in elt. Element "S" is a time evolution plot of the normalized sensitivities. Element "P" is a time evolution plot of the absolute value of the normalized sensitivities. Element "A" is a heatmap of the aliasing scores. Element "T" is like "A", but also showing the aliasing scores as text.

See also

Author

Martijn van Noort